requirements.txt
black==25.12.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
defusedxml==0.7.1
filelock==3.20.3
fsspec==2026.1.0
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
httplib2==0.31.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
packaging==25.0
pandas==2.3.3
pathspec==1.0.3
platformdirs==4.5.1
proto-plus==1.27.0
protobuf==6.33.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
youtube-transcript-api==0.6.2
yt-dlp==2025.12.8
